### 2024-01-29 #6303 @editorialbot
[PRE REVIEW]: whalestrike: An R package for simulating ship strikes on whales
**Submitting author:** @dankelley (<a href="http://orcid.org/0000-0001-7808-5911">Dan E. Kelley</a>)
**Repository:** https://github.com/dankelley/whalestrike
**Branch with paper.md** (empty if default branch): main
**Version:** 0.4.2
