### 2024-01-29 #6303 @kthyng
Hi @dankelley and thanks for your submission! We have a backlog of submissions so I will add this to our waitlist. In the meantime, please suggest 5 reviewers from the database listed above or your own (non-conflicted) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them. Thanks for your patience.

You can name your reviewer suggestions here — make sure they're not [conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?highlight=conflict#joss-conflict-of-interest-policy). Also I see your docs are in the form of a youtube video — I haven't seen that before! We'll see what the reviewers think. I would normally expect a bit more in text/user manual form. And indeed sometimes it takes awhile to find reviewers... Please take a look at the missing DOIs above, and link in your related paper here.
