### 2024-01-30 #6303 @dankelley
@kthyng another good reviewer would be

3. hansenjohnson

and that could be a substitute for clayton33, if the latter is considered conflicted.  But none of this work is funded or part of formal research programs, so I don't feel that clayton33 is conflicted because of contributions made to oce over a decade ago.  (The recent paper coauthored with clayton33 was a way to let folks know about a package that has been on CRAN since 2009.)

Also, the main docs for the `whalestrike` package are conventional: a README, a vignette, and a slew of R manpages.  I just added the video the other day.  It's a new thing I'm doing a lot with students and colleagues.  In this case, it deals with just a single function, buy likely the one (an R shiny app) that newcomers will try first.  I made that app to help in discussions with policy makers.

I hope this covers things.  Github is no longer sending me notifications (no idea why) but I'll check every day or so.
