+++ joss/paper.bib
@@ -28,6 +28,7 @@
 	author = {Vanderlaan, A. S. M. and Taggart, C. T. and Serdynska, A. and Kenney, R. D. and Brown, M. W.},
 	year = {2008},
 	pages = {283--297},
+    doi = {10.3354/esr00083},
 }
 
 @misc{transport_canada_protecting_2018,
@@ -111,7 +112,8 @@
 	url = {https://www.narwc.org/uploads/1/1/6/6/116623219/2022reportcardfinal.pdf},
 	publisher = {North Atlantic Right Whale Consortium},
 	author = {Pettis, H. M. and Pace, R. M. and Hamilton, P. K.},
-	year = {2023}
+	year = {2023},
+    doi = {10.1575/1912/29601},
 }
 
 @article{pace_state-space_2017,
@@ -147,6 +149,7 @@
 	urldate = {2024-01-27},
 	author = {{CRAN}},
 	year = {2024},
+	doi = {10.1002/wics.1212},
 }
 
 @article{schulz_winter_2023,
