### 2024-01-30 0708e33
improve README:
1. Add 'active' badge, as recommended by codemetar.
2. Improve phrasing of docs on app2().
3. Provide link to PDF of the MMS paper.  (Wasn't that open-read
   already?  I thought we paid for that.  Oh well.)
