+++ README.md
@@ -6,6 +6,9 @@
 [![GitHub last commit](https://img.shields.io/github/last-commit/dankelley/whalestrike)](https://img.shields.io/github/last-commit/dankelley/whalestrike)
 [![R-CMD-check](https://github.com/dankelley/whalestrike/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/dankelley/whalestrike/actions/workflows/R-CMD-check.yaml)
 [![status](https://joss.theoj.org/papers/570201320eb0182aa487026819021c50/status.svg)](https://joss.theoj.org/papers/570201320eb0182aa487026819021c50)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+
+
 <!-- badges: end -->
 
 ## Purpose
@@ -44,28 +47,34 @@ app2()
 ```
 
 in an R session.  A window will open in your browser.  On the left-hand side is
-a region with controllers.  Only one marked `Ship` will be open at first.  Try
-adjusting the ship speed, to see what happens to the plots.  Then try other
-elements of the `Ship` controller.  Then try closing that controller, opening
-the others and exploring.  The app has a `Help` button that gives more
-information.  The `Code` button is also handy: it presents a dialog box with
-the raw R code that could be used outside the app, to generate the plots shown.
-In other words, `app2()` acts as both a tool and a tutor.  A video
-demonstrating `app2()` is at available on
-[youtube](https://youtu.be/kTMl3nXa5A4).
+a region with controllers, many hidden in sub-panels that can be opened by
+clicking on V-shaped icons.  Before opening those sub-panels, try adjusting the
+ship speed, to see what happens to the plots, especially the one labelled
+`Lethality Index`. Think about whether the tendency of variations to the index
+are in line with your intuition. Next, open the `Ship` sub-panel, to explore
+the result of altering the ship mass.  This sub-panel also has controllers
+specifying the geometry of the impact region, and you ought to explore them,
+also. Continuing to explore the app's controllers ought to give you a good
+indication of what the tool provides. To learn more, try consulting the app's
+documentation, provided (a) as the response to typing `? app2` in the R
+console, (b) as information in a dialog box that opens when the `Help` button
+is clicked, and (c) in a [youtube video](https://youtu.be/kTMl3nXa5A4).
 
 Users who want more control, and who want to deal with the results in numerical
-as opposed to graphical form, should then start exploring the `strike()`
-function.  To learn how it works, type
+as opposed to graphical form, should try clicking the `Code` button in
+`app2()`, to see the R code that runs the simulation outside the app.  The next
+step will be to explore the functions used in that code.  Documentation exists
+for each of these functions e.g. typing
 ```R
 library(whalestrike)
 ?strike
 ```
-in an R session.
+in an R session will provide information on `strike()`, which is a key function
+in the package.
 
 ## References
 
 Kelley, Dan E., James P. Vlasic, and Sean W. Brillant. “Assessing the Lethality of Ship
 Strikes on Whales Using Simple Biophysical Models.” Marine Mammal
-Science, October 12, 2020, mms.12745. (https://doi.org/10.1111/mms.12745).
-
+Science, October 12, 2020, mms.12745. ([journal site](https://doi.org/10.1111/mms.12745); 
[…]
