### 2024-01-30 4a1d82a
does this version update to a GH action help
I see the following from the R-CMD-check action:

The following actions uses node12 which is deprecated and will be forced to run on node16: actions/checkout@v2. For more info: https://github.blog/changelog/2023-06-13-github-actions-all-actions-will-run-on-node16-instead-of-node12-by-default/
