+++ .github/workflows/R-CMD-check.yaml
@@ -17,7 +17,7 @@ jobs:
     runs-on: macOS-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v2
+      - uses: r-lib/actions/setup-r@v2.6.5
       - uses: r-lib/actions/setup-pandoc@v2
       - name: Install dependencies
         run: |
