+++ .github/workflows/R-CMD-check.yaml
@@ -18,7 +18,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: r-lib/actions/setup-r@v2.6.5
-      - uses: r-lib/actions/setup-pandoc@v2
+      - uses: r-lib/actions/setup-pandoc@v2.6.5
       - name: Install dependencies
         run: |
           install.packages(c("remotes", "rcmdcheck"))
