+++ .github/workflows/R-CMD-check.yaml
@@ -17,8 +17,8 @@ jobs:
     runs-on: macOS-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v2.6.5
-      - uses: r-lib/actions/setup-pandoc@v2.6.5
+      - uses: r-lib/actions/setup-r@v4
+      - uses: r-lib/actions/setup-pandoc@v4
       - name: Install dependencies
         run: |
           install.packages(c("remotes", "rcmdcheck"))
