+++ .github/workflows/R-CMD-check.yaml
@@ -16,9 +16,9 @@ jobs:
   R-CMD-check:
     runs-on: macOS-latest
     steps:
-      - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v4
-      - uses: r-lib/actions/setup-pandoc@v4
+      - uses: actions/checkout@v2
+      - uses: r-lib/actions/setup-r@v2
+      - uses: r-lib/actions/setup-pandoc@v2
       - name: Install dependencies
         run: |
           install.packages(c("remotes", "rcmdcheck"))
