### 2024-02-26 #6303 @diazrenata
@hansenjohnson and @alexee-mayette: Hi! Would you be able and interested in reviewing this submission on simulating the dynamics of ship-whale strikes to the Journal of Open Source Software?
