### 2024-02-26 #6303 @hansenjohnson
Hi @diazrenata! I would be more than happy to review this submission. This would be my first review for JOSS, so any documentation or guidance would be much appreciated. Please let me know how to proceed. Thank you!
