### 2024-02-26 #6303 @diazrenata
@hansenjohnson Wonderful - thank you so very much! I should have linked you to this straightaway - here is the JOSS Reviewer Guide: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Please let me know if you have any questions after reading this! And if it looks good to you, I'll go ahead and add you as a reviewer 😄
