### 2024-02-27 #6303 @hansenjohnson
Hi @diazrenata! I checked out the guidelines and would be happy to review. I should disclose that @dankelley and I were both members of the same department until I graduated in April 2022. I subsequently worked on a short term contract in a different department at the same institution from approximately March-August 2023. I do not believe this constitutes a conflict of interest under the JOSS policy, but defer to your judgement. Please let me know what you think and if you need any more information from me. Thanks!
