### 2024-03-05 #6303 @diazrenata
Hi @hansenjohnson! Thank you for letting me know. Since you were a contracter, it's a little bit of a grey area. Which department(s) were you in? Were they closely related to @dankelley's department?
