### 2024-03-14 dankelley/whalestrike#25 @alexee-mayette
Typo in parameters() function description
On the reference page for the [parameters()](https://dankelley.github.io/whalestrike/reference/parameters.html) function, there is a typo in the description of the Ss argument.

**Ss**: This, together with Cs, is used by **_used by_** [shipWaterForce()](https://dankelley.github.io/whalestrike/reference/shipWaterForce.html) to estimate ship drag force.
