### 2024-03-14 dankelley/whalestrike#26 @alexee-mayette
Duplicated text in the shipAreaFromMass() and shipWaterForce() functions description
In both pages for the [shipAreaFromMass()](https://dankelley.github.io/whalestrike/reference/shipAreaFromMass.html) and [shipWaterForce()](https://dankelley.github.io/whalestrike/reference/shipWaterForce.html), the text for the description is duplicated.
