### 2024-03-14 dankelley/whalestrike#27 @alexee-mayette
whaleAreaFromLength() does not work with other whale species
The function whaleAreaFromLength() does not work when using other species than "N. Atl. Right Whale".

```
whaleAreaFromLength(13.7, "Blue Whale", type = "wetted")

Error in whaleAreaFromLength(13.7, "Blue Whale", type = "wetted") :
  unknown species "Blue Whale"; use one of the following: "N. Atl. Right Whale"
```
