### 2024-03-14 #6473 @alexee-mayette
A few comments for the software paper

- If you export the citation for Kelley et al. 2020 from Marine Mammal Science, they use 2021 as the published date.
- Line 18: “critically Endangered”, letter C should be capitalized.
- I think it would be beneficial to highlight at the end of the Statement of need, that the package allows to compute the probability of lethality by whale species, with default morphometric parameters. This is particularly important as vessel strike risk assessments tend to use the equation from [Conn and Silber (2013)](https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/ES13-00004.1#) which aggregated data from all species of whale together.
- Maybe clearly state that there is no similar package like this out in the public at this moment and that it differs from others like the `shipstrike` package (https://rdrr.io/github/ericmkeen/shipstrike/man/).

I also identified two typo issues [here](https://github.com/dankelley/whalestrike/issues/25#issue-2186369413) and [here](https://github.com/dankelley/whalestrike/issues/26#issue-2186372697) from the vignette page. One [function](https://github.com/dankelley/whalestrike/issues/27#issue-2186381335) did not work when inputting other species than the North Atlantic right whale.
