### 2024-03-14 98701aa
address JOSS review by Alexandra Mayette
Wow, what a careful and insightful review this was.
