+++ joss/paper.md
@@ -27,7 +27,7 @@ North Atlantic right whale. Although speed reduction policies have been
 developed and implemented, simple considerations suggest that other factors,
 such as ship mass and prow shape, should also be taken into account. An R
 package called `whalestrike` has been developed to address such issues. It was
-used by @kelley_assessing_2020 in the development of a biomechanically based
+used by @kelley_assessing_2021 in the development of a biomechanically based
 criterion for the lethality of ship strikes, but this was just a starting
 point. The next step, and goal of the present paper, is to introduce the model
 code to a broader community, encouraging its use and development by diverse
@@ -38,7 +38,7 @@ researchers and policy makers.
 Ship collisions pose significant threats to endangered marine mammals
 [@laist_collisions_2001]. The case of the North Atlantic right whale
 (*Eubalaena glacialis*) is of particular concern, since it is considered a
-"critically Endangered" species [@iucn_eubalaena_2020].  The trend is not
+"Critically Endangered" species [@iucn_eubalaena_2020].  The trend is not
 encouraging for this animal, the population of which was $340\pm7$ in 2021
 [@pettis_north_2023], down significantly from $483$ in 2010
 [@pace_state-space_2017].  According to necropsy studies, ship strikes are
@@ -52,18 +52,27 @@ policy changes, including both static and dynamic zones of speed restriction
 the success of speed restrictions by counting dead or injured whales, given the
 low numbers alive today. In addition, basic reasoning reveals that speed cannot
 be the only factor. Variations in ship mass, prow shape, etc. should also be
-considered, making for a multifactorial problem that requires even more data to
-achieve statistical reliability.
+considered, along with whale morphometric parameters, making for a
+multifactorial problem that requires even more data to achieve statistical
+reliability.
 
-With this in mind, @kelley_assessing_2020 devised a simple numerical model of
+With this in mind, @kelley_assessing_2021 devised a simple numerical model of
 the biophysical dynamics of collisions between vessels and whales. This
-involved using published records of whale injury and death to calibrate a
-lethality criterion that depends not just on vessel speed, but also on other
-factors such as vessel mass, prow geometry, etc. The model was expressed in the
-R language, because it is familiar to many marine biologists and because it
+involved using published records of whale injury and death (across multiple
+species) to calibrate a lethality criterion that depends not just on vessel
+speed, but also on other factors such as vessel mass, prow geometry, etc., in
+addition to whale morphometric parameters. The model was expressed in the R
+language, because it is familiar to many marine biologists and because it
 provides a wide scope of statistical tools for followup work. The purpose of
-the present paper is not to recapitulate the results of @kelley_assessing_2020,
-but rather to introduce readers to the code.
+the present paper is not to recapitulate the results of @kelley_assessing_2021,
+but rather to introduce readers to the `whalestrike` code.
+
+To the author's knowledge, there are no other R packages that address the topic
+of collision mechanics and whale lethality in this way.  Readers interested in
+the related topic of the probability of collision given shipping patterns and
+whale distributions might find it helpful to start with R packages named
+`whalemap` (@johnson_whalemap_2021) and `shipstrike`
+(@keen_ericmkeenshipstrike_2023).
 
 # Model formulation
 
@@ -85,7 +94,7 @@ with both extension and compression forces, while only compression is
 considered for interior layers. Lacking reliable information on failure limits
 for these biomaterials, critical values for stresses are posited in the context
     of published results of the damage experienced in documented ship strikes.
-    @kelley_assessing_2020 provide more information on the methodologies
+    @kelley_assessing_2021 provide more information on the methodologies
     and issues involved.
 
 The model is simple, with acceleration and forces being linked via Newton's
[…]
