+++ joss/paper.bib
@@ -167,7 +167,7 @@
     doi = {10.1029/2022JC019197},
 }
 
-@article{kelley_assessing_2020,
+@article{kelley_assessing_2021,
 	title = {Assessing the lethality of ship strikes on whales using simple biophysical models},
 	volume = {37},
 	url = {https://doi.org/10.1111%2Fmms.12745},
@@ -175,8 +175,8 @@
 	number = {1},
 	journal = {Marine Mammal Science},
 	author = {Dan E. Kelley and James P. Vlasic and Sean W. Brillant},
-	month = oct,
-	year = {2020},
+	month = jan,
+	year = {2021},
 	pages = {251--267},
 }
 
@@ -187,4 +187,30 @@
 	author = {Dan E. Kelley},
 	month = jan,
 	year = {2024},
+}
+
+    
+@misc{keen_ericmkeenshipstrike_2023,
+	title = {ericmkeen/shipstrike: {Utilities} for spatially and temporally explicit predictions of whale-ship collisions},
+	url = {https://github.com/ericmkeen/shipstrike},
+    date = 2023,
+	urldate = {2024-03-14},
+	author = {Keen, Eric},
+	keywords = {whale},
+}
+
+@article{johnson_whalemap_2021,
+	title = {{WhaleMap}: a tool to collate and display whale survey results in near real-time},
+	volume = {6},
+	issn = {2475-9066},
+	shorttitle = {{WhaleMap}},
+	url = {https://joss.theoj.org/papers/10.21105/joss.03094},
+	doi = {10.21105/joss.03094},
+	number = {62},
+	urldate = {2024-03-14},
+	journal = {Journal of Open Source Software},
+	author = {Johnson, Hansen and Morrison, Daniel and Taggart, Christopher},
+	month = jun,
+	year = {2021},
+	pages = {3094},
 }
