### 2024-03-14 dankelley/whalestrike#27 @dankelley
This issue has been addressed in the "main" branch of
 https://github.com/dankelley/whalestrike
in commit  98701aa997784dfe99fd92908c9a466464f5eb82, by altering the function description to let users know that the package only handles one species, in this version.
