### 2024-03-14 #6473 @dankelley
@alexee-mayette thanks *so* much for your rapid and very insightful reading of not just the article, but also the code.  I have addressed the three issues, as well as your comments on the paper, in the "main" branch of the repository at https://github.com/dankelley/whalestrike, in commit  98701aa997784dfe99fd92908c9a466464f5eb82.

I think the paper will auto-regenerate, but I'll check on that after waiting a while to see if it happens automatically.
