### 2024-03-14 a249075
codemeta update; spellcheck update; url update (but...)
... the urlchecker complains about the journal website for the official
paper, and also for the researchgate site.  Perhaps neither is set up in
a way that CRAN likes. This seems to be not uncommon.
