+++ joss/paper.bib
@@ -167,19 +167,6 @@
     doi = {10.1029/2022JC019197},
 }
 
-@article{kelley_assessing_2021,
-	title = {Assessing the lethality of ship strikes on whales using simple biophysical models},
-	volume = {37},
-	url = {https://doi.org/10.1111%2Fmms.12745},
-	doi = {10.1111/mms.12745},
-	number = {1},
-	journal = {Marine Mammal Science},
-	author = {Dan E. Kelley and James P. Vlasic and Sean W. Brillant},
-	month = jan,
-	year = {2021},
-	pages = {251--267},
-}
-
 @misc{dan_kelley_using_2024,
 	title = {Using app2()  of the whalestrike {R} package},
 	url = {https://www.youtube.com/watch?v=kTMl3nXa5A4},
@@ -213,4 +200,26 @@
 	month = jun,
 	year = {2021},
 	pages = {3094},
+}
+
+
+
+
+
+
+
+@article{kelley_assessing_2021,
+	title = {Assessing the lethality of ship strikes on whales using simple biophysical models},
+	volume = {37},
+	issn = {0824-0469, 1748-7692},
+	url = {https://onlinelibrary.wiley.com/doi/10.1111/mms.12745},
+	doi = {10.1111/mms.12745},
+	language = {en},
+	number = {1},
+	urldate = {2024-03-14},
+	journal = {Marine Mammal Science},
+	author = {Kelley, Dan E. and Vlasic, James P. and Brillant, Sean W.},
+	month = jan,
+	year = {2021},
+	pages = {251--267},
 }
