+++ man/app2.Rd
@@ -38,7 +38,7 @@ sense?  Continue this process, exploring all the panes. It is hoped
 that a half hour of such exploration will let users start to
 investigate practical applications.  For more about how the simulations
 are carried out, as well as comments on some applications that
-may be of interest, please consult Kelley et al. (2020).
+may be of interest, please consult Kelley et al. (2021).
 
 More information on \code{\link[=app2]{app2()}} in video form on
 \href{https://youtu.be/kTMl3nXa5A4}{youtube}.
@@ -46,7 +46,7 @@ More information on \code{\link[=app2]{app2()}} in video form on
 \references{
 Kelley, Dan E., James P. Vlasic, and Sean W. Brillant. "Assessing the Lethality
 of Ship Strikes on Whales Using Simple Biophysical Models." Marine Mammal
-Science, October 12, 2020, mms.12745. \doi{10.1111/mms.12745}.
+Science, 37(1), 2021. \doi{10.1111/mms.12745}.
 }
 \seealso{
 Other interactive apps: 
