+++ README.md
@@ -79,4 +79,4 @@ Kelley, Dan E., James P. Vlasic, and Sean W. Brillant. "Assessing the Lethality
 of Ship Strikes on Whales Using Simple Biophysical Models." Marine Mammal
 Science 37, no. 1 (January 2021): 251–67. https://doi.org/10.1111/mms.12745.
 ([journal site](https://doi.org/10.1111/mms.12745); [ResearchGate
-Preprint](https://www.researchgate.net/publication/344748816_Assessing_the_lethality_of_ship_strikes_on_whales_using_simple_biophysical_models#fullTextFileContent))
+Preprint](https://www.researchgate.net/publication/344748816_Assessing_the_lethality_of_ship_strikes_on_whales_using_simple_biophysical_models))
