### 2024-04-23 #6473 @diazrenata
@hansenjohnson, just checking in on your review for `whalestrike`!
