### 2024-05-13 #6473 @hansenjohnson
My sincere apologies for my very delayed review. I have had a chance to thoroughly inspect the software and associated documentation. It appears to be very well organized, high-quality work. I have updated the checklist above accordingly. I flagged a few very minor issues and considerations during my review:

Minor issues

- Link to youtube in the shiny app help popup is not displaying as a link
- “Auto-refined grid to capture acceleration peak” message overwhelms the app screen when clicking and dragging the slider.
- Typo in line 66 of the manuscript – remove parenthesis around reference

Suggestions

- It would be helpful to have a small info button next to each of the plot options that could be clicked on to provide a very brief description of what each plot is showing. This was described verbally in the youtube video and I found it very helpful.
- I understand that the whale selector dropdown changes the model used to calculate whale mass. I suggest it should also change preset values for whale length and the various tissue thicknesses. This would make it much more convenient to make realistic comparisons between species, without requiring the user to look up those details.
- On a similar note, it would be very helpful to have a dropdown menu with preset values for different types of vessels (e.g., container, cruise ship, car ferry, recreational fishing, offshore fishing, sailing). That also make it much more convenient for the user to make realistic comparisons.
- Is there any interest in or plan to deploy this app on a shiny server so that users can interact with the GUI without any knowledge of R? I suspect that, combined with the preset suggestions above, would make it much more approachable to a wide range of researchers who do not have a working knowledge of R.
