### 2024-05-13 #6473 @dankelley
Thanks very much for the review, @hansenjohnson.  I'll take a look at these things in the next few days.  The last three of your suggestions are a little bit outside the scope of this journal submission, I think, because they will require more work.  When that can be done, I'm not too sure.

Re suggestions 2 and 3: there are no students to help with this, so addressing these (very fine) suggestions means me scouring the literature to try to find details of whale and vessel types.  That's quite a lot of work, and it really ought to be done by a student, I think, because projects like that let a student contribute meaningfully to a larger goal, whilst doing background reading that can help more broadly.  I don't have such a student upon whom I can call, so these two suggestions are something that I hope could be done in collaboration with the community.  (That's basically why I submitted this manuscript to this journal, in fact.  I'm hoping that somebody working on species X, vessel Y, will offer data to help with that circumstance, aiding not just in their own work, but for others as well.

Re suggestion 4: finances are an issue. We do not have computing resources in my department, and the university resources require a funding stream that I do not have for this work.  Sometimes people will use a machine funded to work on topic X to host computation for topic Y, but I don't like to ask people to do that.  (My work on university Senate makes me a bit too aware of rules, I guess.)

The delay was not a problem at all -- as you know, my main work is in an entirely different area -- and thanks again for these really great suggestions.

Dan.
