### 2024-05-13 3ea006d
address hansenjohnson minor issues 1, 2 and 3
I plan to address suggestion 1 within a day or two.
