### 2024-05-13 #6473 @dankelley
commit 3ea006dc09849c0b9829d11e25cefb99ece4d45f
Author: dankelley <kelley.dan@gmail.com>
Date:   Mon May 13 11:23:28 2024 -0300

    address hansenjohnson minor issues 1, 2 and 3

    I plan to address suggestion 1 within a day or two.
