+++ R/app2.R
@@ -74,8 +74,10 @@ simulation under view at any given time.
 #' @param debug logical value indicating whether to print output to
 #' the R console as the computation is done.
 #'
-#' @importFrom bslib accordion accordion_panel card sidebar
+#' @importFrom bslib accordion accordion_panel card card_header sidebar tooltip
+#'
 #' @importFrom shiny actionButton checkboxGroupInput h6 observeEvent plotOutput selectInput shinyApp showModal sliderInput stopApp
+#'
 #' @export
 #'
 #' @family interactive apps
@@ -99,18 +101,36 @@ app2 <- function(debug = FALSE) {
     }
     ui <- bslib::page_sidebar(
         # title = "app2", # a waste of space (the user launched this, and can do help)
-        shiny::tags$script(paste0("$(document).on(\"keypress\", function (e) {",
+        shiny::tags$script(paste0(
+            "$(document).on(\"keypress\", function (e) {",
             "Shiny.onInputChange(\"keypress\", e.which);",
             "Shiny.onInputChange(\"keypressTrigger\", Math.random());",
-            "});")),
+            "});"
+        )),
         sidebar = bslib::sidebar(
             # title = "Controls", # a waste of space (who could not figure this out?)
+            tooltip(
+                "\u24D8",
+                paste(
+                    "The three 'accordion' menus below are used to select the",
+                    "desired plot types, and to set the characteristics of the",
+                    "whale and the ship that collides with it."
+                )
+            ),
             width = 280, # default, 250, too narrow for one of the plot types
             bslib::accordion(
                 open = FALSE,
                 multiple = TRUE,
                 accordion_panel(
                     "plot",
+                    tooltip(
+                        "\u24D8",
+                        paste(
+                            "The checkboxes select the plots to be displayed.",
+                            "To learn more about the choices, try typing",
+                            "'?whalestrike::plot.strike' in an R session."
+                        )
+                    ),
                     shiny::checkboxGroupInput("plot_panels", "",
                         choices = c(
                             "location",
@@ -129,6 +149,14 @@ app2 <- function(debug = FALSE) {
                 ),
                 accordion_panel(
                     "whale",
+                    tooltip(
+                        "\u24D8",
+                        paste(
+                            "These sliders set the whale characteristics.",
+                            "The default values are meant to represent",
+                            "an adult Right Whale."
+                        )
+                    ),
                     shiny::selectInput("species", "Species",
                         choices = c(
                             "N. Atl. Right", "Blue", "Bryde", "Fin", "Gray", "Humpback", "Minke",
@@ -159,7 +187,19 @@ app2 <- function(debug = FALSE) {
                 ),
                 accordion_panel(
                     "ship",
-                    shiny::sliderInput("ms", shiny::h6("Mass [tonne]"),
+                    tooltip(
+                        "\u24D8",
+                        paste(
+                            "These sliders set the mass of the ship and the",
+                            "geometry of the impact zone. For the latter,",
+                            "consider the area a few decimetres aft of the first",
+                            "contact point. The default values correspond to a small",
+                            "fishing vessel with a prow shape similar to that of a",
+                            "Cape Islander."
+                        )
+                    ),
+                    shiny::sliderInput("ms",
+                        shiny::h6("Mass [tonne]"),
                         ticks = FALSE,
                         min = 10, max = 500, value = 45, step = 1
                     ),
@@ -171,28 +211,57 @@ app2 <- function(debug = FALSE) {
                         ticks = FALSE,
                         min = 0.1, max = 2, value = 1.15, step = 0.05
[…]
