+++ README.md
@@ -13,22 +13,24 @@
 
 ## Purpose
 
-This package provides tools for simulating the collisions of ships with whales,
-using a simplified dynamical structure involving point masses separated by
-compressible materials. Along with functions for computations of forces,
-deformations, accelerations, etc., the package provides a easy-to-use GUI tool
-that makes it easy to set up some common simulation scenarios, and to display
-the results in graphical form.
+This package provides tools for simulating the collisions of ships
+with whales, using a simplified dynamical structure involving point
+masses separated by compressible materials. Along with functions for
+computations of forces, deformations, accelerations, etc., the package
+provides a easy-to-use GUI tool that makes it easy to set up some
+common simulation scenarios, and to display the results in graphical
+form.
 
-To learn more about the scientific background, and to see the results of some
-detailed computations placed in the context of a database of observed strikes,
-see Kelley et al. (2021).
+To learn more about the scientific background, and to see the results
+of some detailed computations placed in the context of a database of
+observed strikes, see Kelley et al. (2021). For more information on
+the coding, see Kelley (2024).
 
 ## Installation
 
-The package is not yet available on CRAN, and must be installed from source.
-This can be done either by downloading the source and building it locally, or
-by the simpler method of typing
+The package is not yet available on CRAN, and must be installed from
+source. This can be done either by downloading the source and building
+it locally, or by the simpler method of typing
 ```R
 # install.package("remotes")
 remotes::install_github("dankelley/whalestrike", ref="main")
@@ -46,37 +48,47 @@ library(whalestrike)
 app2()
 ```
 
-in an R session.  A window will open in your browser.  On the left-hand side is
-a region with controllers, many hidden in sub-panels that can be opened by
-clicking on V-shaped icons.  Before opening those sub-panels, try adjusting the
-ship speed, to see what happens to the plots, especially the one labelled
-`Lethality Index`. Think about whether the tendency of variations to the index
-are in line with your intuition. Next, open the `Ship` sub-panel, to explore
-the result of altering the ship mass.  This sub-panel also has controllers
-specifying the geometry of the impact region, and you ought to explore them,
-also. Continuing to explore the app's controllers ought to give you a good
-indication of what the tool provides. To learn more, try consulting the app's
-documentation, provided (a) as the response to typing `? app2` in the R
-console, (b) as information in a dialog box that opens when the `Help` button
-is clicked, and (c) in a [youtube video](https://youtu.be/kTMl3nXa5A4).
+in an R session.  A window will open in your browser.  On the
+left-hand side is a region with controllers, many hidden in sub-panels
+that can be opened by clicking on V-shaped icons.  Before opening
+those sub-panels, try adjusting the ship speed, to see what happens to
+the plots, especially the one labelled `Lethality Index`. Think about
+whether the tendency of variations to the index are in line with your
+intuition. Next, open the `Ship` sub-panel, to explore the result of
+altering the ship mass.  This sub-panel also has controllers
+specifying the geometry of the impact region, and you ought to explore
+them, also. Continuing to explore the app's controllers ought to give
+you a good indication of what the tool provides. To learn more, try
+consulting the app's documentation, provided (a) as the response to
+typing `? app2` in the R console, (b) as information in a dialog box
+that opens when the `Help` button is clicked, and (c) in a [youtube
+video](https://youtu.be/kTMl3nXa5A4) and a brief [followup youtube
+video](https://youtu.be/f8nHGikb9ug) that illustrates an alteration
[…]

[…]
