### 2024-05-13 #6473 @dankelley
As noted below, I have addressed the issues raised by @hansenjohnson, and also the first of the suggestions.  I also made a 1.5 minute video illustrating the use of the popup tooltip items (it is listed in the README now, which can be seen at https://dankelley.github.io/whalestrike but you can also get it directly at https://www.youtube.com/watch?v=f8nHGikb9ug&feature=youtu.be)

I think this satisfies the issues that have been raised.  I will of course keep in mind suggestions 2, 3 and 4 mentioned in @hansenjohnson's review, for future work.  And, as noted above, part of my reason for writing this JOSS manuscript was to let users know that I'll welcome additions to the package.

@diazrenata -- are there any other items I ought to be taking care of at this stage?

commit 01468b522049f5c9338ff73409c9b15283efef38 (HEAD -> main, origin/main, origin/HEAD)
Author: dankelley <kelley.dan@gmail.com>
Date:   Mon May 13 16:00:41 2024 -0300

    address concerns raised in Johnson's JOSS review
