+++ joss/paper.bib
@@ -63,13 +63,6 @@
 	year = {2010},
 }
 
-@techreport{kite-powell_modeling_2007,
-	title = {Modeling the effect of vessel speed on {Right} {Whale} ship strike risk},
-	shorttitle = {Modeling the effect of vessel speed on {Right} {Whale} ship strike risk},
-	author = {Kite-Powell, H. K. and Knowlton, A. and Brown, M. W.},
-	year = {2007},
-}
-
 @article{laist_collisions_2001,
 	title = {Collisions between ships and whales},
 	volume = {17},
@@ -168,7 +161,7 @@
 }
 
 @misc{dan_kelley_using_2024,
-	title = {Using app2()  of the whalestrike {R} package},
+	title = {Using app2() of the whalestrike {R} package},
 	url = {https://www.youtube.com/watch?v=kTMl3nXa5A4},
 	urldate = {2024-01-27},
 	author = {Dan E. Kelley},
