### 2024-05-28 #6473 @diazrenata
@dankelley Could you fill in the missing DOIs for the references named above? I know several of them won't have DOIs, and that's fine, but at least this one should: https://doi.org/10.1638/2006-0057.1
