### 2024-05-28 bf214fb
add a DOI (thanks to JOSS editor for pointing this out)
