+++ joss/paper.bib
@@ -1,198 +1,226 @@
 @mastersthesis{raymond_development_2007,
-	title = {Development of a numerical model to predict impact forces on a {North} {Atlantic} {Right} {Whale} during collision with a vessel},
-	url = {https://scholars.unh.edu/thesis/309},
+    title = {Development of a numerical model to predict impact forces on a {
+             North} {Atlantic} {Right} {Whale} during collision with a vessel},
+    url = {https://scholars.unh.edu/thesis/309},
     school = {University of New Hampshire},
-	author = {Raymond, J. J.},
-	year = {2007},
+    author = {Raymond, J. J.},
+    year = {2007},
 }
 
 @article{conn_vessel_2013,
-	title = {Vessel speed restrictions reduce risk of collision‐related mortality for {North} {Atlantic} right whales},
-	volume = {4},
-	issn = {2150-8925},
-	url = {https://doi.org/10.1890/ES13-00004.1},
-	doi = {10.1890/ES13-00004.1},
-	number = {4},
-	urldate = {2018-06-18},
-	journal = {Ecosphere},
-	author = {Conn, P. B. and Silber, G. K.},
-	year = {2013},
-	pages = {art43},
+    title = {Vessel speed restrictions reduce risk of collision‐related
+             mortality for {North} {Atlantic} right whales},
+    volume = {4},
+    issn = {2150-8925},
+    url = {https://doi.org/10.1890/ES13-00004.1},
+    doi = {10.1890/ES13-00004.1},
+    number = {4},
+    urldate = {2018-06-18},
+    journal = {Ecosphere},
+    author = {Conn, P. B. and Silber, G. K.},
+    year = {2013},
+    pages = {art43},
 }
 
 @article{vanderlaan_reducing_2008,
-	title = {Reducing the risk of lethal encounters: vessels and right whales in the {Bay} of {Fundy} and on the {Scotian} {Shelf}},
-	volume = {4},
-	shorttitle = {Reducing the risk of lethal encounters: vessels and right whales in the {Bay} of {Fundy} and on the {Scotian} {Shelf}},
-	journal = {Endangered Species Research},
-	author = {Vanderlaan, A. S. M. and Taggart, C. T. and Serdynska, A. and Kenney, R. D. and Brown, M. W.},
-	year = {2008},
-	pages = {283--297},
+    title = {Reducing the risk of lethal encounters: vessels and right whales in
+             the {Bay} of {Fundy} and on the {Scotian} {Shelf}},
+    volume = {4},
+    shorttitle = {Reducing the risk of lethal encounters: vessels and right
+                  whales in the {Bay} of {Fundy} and on the {Scotian} {Shelf}},
+    journal = {Endangered Species Research},
+    author = {Vanderlaan, A. S. M. and Taggart, C. T. and Serdynska, A. and
+              Kenney, R. D. and Brown, M. W.},
+    year = {2008},
+    pages = {283--297},
     doi = {10.3354/esr00083},
 }
 
 @misc{transport_canada_protecting_2018,
-	title = {Protecting {North} {Atlantic} right whales from ship strikes in the {Gulf} of {St}. {Lawrence}.},
-	shorttitle = {Protecting {North} {Atlantic} right whales from ship strikes in the {Gulf} of {St}. {Lawrence}.},
-	url = {www.tc.gc.ca/en/services/marine/navigation-marine-conditions/protecting-north-atlantic-right-whales-ship-strikes-gulf-st-lawrence.html},
-	urldate = {2018-08-25},
-	author = {{Transport Canada}},
[…]
