### 2024-05-28 #6473 @dankelley
@diazrenata thanks for noting that missing DOI.  The others do not have DOIs.  I bumped the version number and rebuilt the online docs.  Local tests are ok for me.  So are the remote github-action tests of the package (these tests do not touch the `joss/` directory or its contents, but I like to check things).

Will you let me know when it's OK for me to get a zenodo entry and post it?  (I made one before adding this DOI, and will try to update it now.  I'm a bit rusty on using zenodo.)
