### 2024-05-28 #6473 @dankelley
Oh, yes.  It said updates were not possible but I had bumped the version number and so it worked.  Somehow I got myself confused with the zenodo intterface.

The zenodo record id https://doi.org/10.5281/zenodo.11372537
