### 2024-05-28 #6473 @dankelley
I should be the sole author of this JOSS manuscript, which is describing the software.

There were 3 authors on the science paper in Marine Mammal Science: me, Sean Brillant and James Vlasic.  Sean and I, who were supervising James as a summer student, a few years ago, had agreed that the best plan was for me to single-author this JOSS manuscript about the package, given that I had devised the theoretical model and written the code to actualize it.
