### 2024-05-29 #6473 @kthyng
@dankelley Sorry if I missed this in the discussion above but I see a few versions running around: v2 on Zenodo and v0.5.0 here. Which do you want associated with this JOSS paper? I'd like them to be consistent, and usually you would also make a release of your package associated with the version, too.

Also we usually have the title of the Zenodo archive exactly match the JOSS paper, as in "whalestrike: An R package for simulating ship strikes on whales", though this is a request not a requirement to make the JOSS archive self-consistent. This requires a change in the Zenodo metadata.
