### 2024-05-29 #6473 @dankelley
The DOI for the proper version (i.e. 0.5.0) is [10.5281/zenodo.11372537](https://doi.org/10.5281/zenodo.11372537)
May 28, 2024
