### 2024-05-29 #6473 @kthyng
@dankelley Great, everything looks good.

> The package is not formally released, in the sense of an R package on CRAN. For now, users install it as instructed on page 2 of the JOSS manuscript.

That's fine to wait. In terms of a release I was referring to how you have a tag here that is pretty old: https://github.com/dankelley/whalestrike/tags and usually at this point you'd make another tag to match your v0.5.0. But you don't have to.
