### 2024-05-29 #6473 @kthyng
Congratulations on your new publication @dankelley! Many thanks to @diazrenata (and special shoutout for handling your first JOSS submission all the way through!) and to reviewers @alexee-mayette and @hansenjohnson for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
